Port `load_nic`
